requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
chromedriver-binary==110.0.5481.30.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.4
decorator==5.1.1
Django==4.1.4
django-environ==0.9.0
django_debug_toolbar==3.8.1
exceptiongroup==1.1.0
gunicorn==20.1.0
h11==0.14.0
humanize==4.4.0
idna==3.4
install==1.3.5
kombu==5.2.4
lxml==4.9.2
outcome==1.2.0
Pillow==9.3.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.7
redis==4.4.2
requests==2.28.1
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tornado==6.2
transaction==3.0.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
wsproto==1.2.0
zope.interface==5.5.2
